pyproject.toml
[project]
name = "srag-analytics"
version = "0.1.0"
description = "AI-powered SRAG analytics with LangGraph agents"
requires-python = ">=3.11,<4.0"
dependencies = [
    # LangChain Core
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",

    # LangGraph (latest)
    "langgraph>=0.6.8",
    "langgraph-checkpoint-postgres>=2.0.0",

    # LLM Providers
    "langchain-openai>=0.3.34",

    # Tools & Integrations
    "tavily-python>=0.5.0",

    # Database & Vector Store
    "psycopg[binary,pool]>=3.2.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # API & Web Framework
    "fastapi>=0.115.0",
    "langserve[all]>=0.3.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",

    # Frontend
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "pandas>=2.2.0",

    # PDF Processing
    "pypdf>=5.0.0",
    "pdfplumber>=0.11.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "ipython>=8.28.0",
    "jupyter>=1.1.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "frontend*"]
exclude = ["data*", "infra*", "notebooks*", "docs*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
